Correções

+++ 1P/PEC/24-02-S3/03 - tarefa 1 - runcodes/runcodes/r-t1-q4 - resultado expressão.py
@@ -1,2 +1,2 @@
-x = int(input("").strip())
+x = float(input("").strip())
 print(f'{9 * x - 4 * x + 10}')
